Code cleanup: Unused vars, parameters, etc.
- Made several function parameters unnamed. As the C++ Core Guidelines
  say, "Unused parameters should be unnamed."
- Where the names were part of a signature, added at the top of the
  function:  (void)var_name;  // UNUSED

+++ src/effects/Stabilizer.cpp
@@ -136,7 +136,7 @@ bool Stabilizer::LoadStabilizedData(std::string inputFilePath){
     trajectoryData.clear();
 
     // Iterate over all frames of the saved message and assign to the data maps
-    for (size_t i = 0; i < stabilizationMessage.frame_size(); i++) {
+    for (int i = 0; i < stabilizationMessage.frame_size(); i++) {
 
 		// Create stabilization message
         const pb_stabilize::Frame& pbFrameData = stabilizationMessage.frame(i);
